Handle EOFError for REST API prompt under Docker and move logic into 'src/' folder
